Added New Lookup Tab
Added a new tab that allows for looking up individual shuttles
This required a switch from PagerAdapter to FragmentPagerAdapter, which
made it necessary to create FragmentListView, FragmentShuttleLookup,
and FragmentWeekView to display the tabs.
For simplicity, an interface was added to ConstraintDatabase which
handled listening for database updates.
Finally, a new broadcast type was implemented to allow an intent to
request the arrival time of a specific stop (for
FragmentShuttleLookup’s functionallity)